+++ source/system_calls/write.c
@@ -1,6 +1,6 @@
 #include <liblinux/system_call.h>
 #include <liblinux/system_calls/write.h>
 
-long write(unsigned int file_descriptor, char * buffer, size_t count) {
+ssize_t write(unsigned int file_descriptor, char * buffer, size_t count) {
     return system_call(1, file_descriptor, (long) buffer, count);
 }
